Use floating point calculations for column sizes to avoid accumulating rounding errors
Test Plan:
Starting and quitting kmail multiple times used to make the first column grow
every time.

Reviewers: mlaurent

Reviewed By: mlaurent

Subscribers: #kde_pim

Differential Revision: https://phabricator.kde.org/D5763

+++ messagelist/src/core/theme.cpp
@@ -555,7 +555,7 @@ bool Theme::Row::load(QDataStream &stream, int themeVersion)
     return true;
 }
 
-Theme::Column::SharedRuntimeData::SharedRuntimeData(bool currentlyVisible, int currentWidth)
+Theme::Column::SharedRuntimeData::SharedRuntimeData(bool currentlyVisible, double currentWidth)
     : mReferences(0), mCurrentlyVisible(currentlyVisible), mCurrentWidth(currentWidth)
 {
 }
@@ -591,12 +591,12 @@ void Theme::Column::SharedRuntimeData::setCurrentlyVisible(bool visible)
     mCurrentlyVisible = visible;
 }
 
-int Theme::Column::SharedRuntimeData::currentWidth() const
+double Theme::Column::SharedRuntimeData::currentWidth() const
 {
     return mCurrentWidth;
 }
 
-void Theme::Column::SharedRuntimeData::setCurrentWidth(int currentWidth)
+void Theme::Column::SharedRuntimeData::setCurrentWidth(double currentWidth)
 {
     mCurrentWidth = currentWidth;
 }
@@ -729,12 +729,12 @@ void Theme::Column::setCurrentlyVisible(bool currentlyVisible)
     mSharedRuntimeData->setCurrentlyVisible(currentlyVisible);
 }
 
-int Theme::Column::currentWidth() const
+double Theme::Column::currentWidth() const
 {
     return mSharedRuntimeData->currentWidth();
 }
 
-void Theme::Column::setCurrentWidth(int currentWidth)
+void Theme::Column::setCurrentWidth(double currentWidth)
 {
     mSharedRuntimeData->setCurrentWidth(currentWidth);
 }
